Move class index to index subroute
