+++ app/controllers/project-version/module/index.js
@@ -1,0 +1,3 @@
+import ClassIndexController from "../class/index";
+
+export default ClassIndexController.extend();
