+++ app/controllers/project-version/namespace/index.js
@@ -1,0 +1,3 @@
+import ClassIndexController from "../class/index";
+
+export default ClassIndexController.extend();
