+++ app/routes/project-version/module/index.js
@@ -1,0 +1,5 @@
+import Ember from 'ember';
+
+export default Ember.Route.extend({
+  templateName: 'class-index'
+});
