+++ app/styles/components/_tabbed-layout.scss
@@ -4,7 +4,7 @@
   border-radius: 5px;
   overflow: hidden;
 
-  .api-index-filter {
+  .api-index-filter, section {
     padding: 0 1em;
   }
